+++ packages/shared/src/wallet/index.ts
@@ -7,7 +7,10 @@ export {
 } from '../types/wallet'
 export { useActiveSuiAddress } from './hooks/useActiveSuiAddress'
 export { useAddressAliases } from './hooks/useAddressAliases'
-export type { AddressAliasesInfo } from './hooks/useAddressAliases.config'
+export {
+  ADDRESS_ALIAS_MODULE,
+  type AddressAliasesInfo,
+} from './hooks/useAddressAliases.config'
 export { useBalance } from './hooks/useBalance'
 export { useSendToken } from './hooks/useSendToken'
 export { useTransactionHistory } from './hooks/useTransactionHistory'