+++ apps/extension/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @evevault/extension
 
+## 0.0.14
+
+### Patch Changes
+
+- address aliasing
+- Updated dependencies
+  - @evevault/shared@0.0.14
+
 ## 0.0.13
 
 ### Patch Changes